GH Action requiring multiple internal approvers
This action is meant to replace the
[workflow](https://github.com/abcxyz/pkg/blob/main/.github/workflows/multi-approvers.yml)
of the same name. A GH Action had to be used instead of a Workflow to facilitate
passing tokens from abcxyz/github-token-minter (or any other action that
produces a GH token). This is because of the security ramification of passing
secret tokens between GH Workflow jobs.

**Deprecating the multi-approvers.yml workflow**

The legacy multi-approvers.yml workflow will be deprecated as a
follow-up to this PR. We will first need to migrate any users off of the
legacy workflow before deletion.

**Testing**

Most code is tested via automated tests. Run `npm run test` to run the
tests. Testing uses HTTP mocks using nock to mock the GH REST API.

This action has also been manually tested using test GH repos.

**Follow-up items**

* Deprecate multi-approvers.yml workflow.
* Create generic GH action to re-run workflows on PR reviews (this is
  also needed by want-lgtm and guardian).
* Add tests to presubmit CI/CD workflow.
* Add build to CI/CD workflow.

An issue will be created for each of these.