lint: javascript (#69)
Add a new action for linting javascript/typescript/node based code.

* Runs on every package.json in the repo. Shards to one vm per
package.json
* Runs `npm run lint` if a lint command is defined in the package.json
* Otherwise runs `npx eslint .` if an eslint config is present.
* Otherwise does nothing and instead prints a notice to add a "lint"
command to your package.json or add an eslint config.
* NOTE: I considered using [gts](https://github.com/google/gts) - but it
just adds a `lint` command to your package.json and is not CI friendly
as it generates a lot of source code if you haven't used it before in a
repo.
* TODO: lint standalone js

Tested on github-token-minter by manually adding:
https://github.com/abcxyz/github-token-minter/pull/207