Add allowlist feature for multi-approvers (#74)
Add an allowlist of GitHub user IDs that won't require multi-approvers
when raising a PR.

This add a new input `allowlisted-user-ids` via action, which is an
allowlist of trusted GitHub user IDs. It will be read and parsed into a
Set. `MultiApproversAction` will check if the user in on the allowlist.
If so, it won't trigger multi-approvers requirement like internal users.

Fixes #68