chore: correctly handle file renames in create-pull-request (#93)
Tested in another repo:
https://github.com/porcupine/platform-gcp-infrastructure/actions/runs/15592168989/job/43913706692.
Moved files only show up in the CHANGED files list. They must show up
both in the CHANGED AND DELETED lists in order to properly be handled.

Before making this change modifications that resulted in a `moved file`
were incorrectly committed as `added file` and the old files were not
removed leading to file duplication.

Signed-off-by: Dylan Yurgionas <dcreey@gmail.com>