fix: create-pull-request action using wrong sha as base (#124)
The current code always used `context.sha` as the root sha for the PR
branch. This worked accidentally when creating PRs for the workflow that
the action was running for, but failed when attempting to create PRs in
another repository.

Modified to use the base branch (main) head reference as the root for
the branch.